requirements.txt
# ComplianceAstra Backend - Phase 1
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.18
stripe==11.4.0
reportlab==4.2.5
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
greenlet
aiosqlite==0.20.0
